requirements.txt
Flask==2.0.1
Pillow==8.3.1
pyyaml==6.0.1 